move bookmarks functionality outside components folder
